feat: reset filters when switching accounts
